+++ examples/advanced_chat/chat-stream.ts
