--- examples/simple_chat/main.ts
@@ -1,20 +1,0 @@
-import { TinfoilAI } from "tinfoil";
-
-async function main() {
-  try {
-    const client = new TinfoilAI({
-    });
-
-    const completion = await client.chat.completions.create({
-      messages: [{ role: "user", content: "Hello!" }],
-      model: "llama-free",
-    });
-
-    console.log(completion.choices[0]?.message?.content);
-  } catch (error) {
-    console.error("Error:", error);
-    process.exit(1);
-  }
-}
-
-main();
